WIP: DriveSetup: add "disk image menu"
Allows to register and unregister disk images, as well as backup a
partition to an image or restore an image to a partition.

A new "task view" is placed at the bottom of DriveSetup to track multiple write/save sessions.
This could also be used in the future to display other sorts of "disk tasks".

Change-Id: I4e377b1d521ceb7a91e1a5c0d1bb3d37b24fcd68